demo/pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = "streamlit page for demo"
authors = ["hobbang2 <cattie@pukyong.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
pnumpy = "^2.0.23"
pandas = "2.0"
streamlit = "^1.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
